Move footer to ScreenTestScene
