+++ osu.Game/Tests/Visual/ScreenTestScene.cs
@@ -7,7 +7,9 @@ using osu.Framework.Bindables;
 using osu.Framework.Development;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Cursor;
 using osu.Framework.Logging;
+using osu.Framework.Screens;
 using osu.Framework.Testing;
 using osu.Game.Graphics;
 using osu.Game.Overlays;
@@ -32,7 +34,7 @@ namespace osu.Game.Tests.Visual
         protected DialogOverlay DialogOverlay { get; private set; }
 
         [Cached]
-        private ScreenFooter footer;
+        protected ScreenFooter Footer { get; private set; }
 
         protected ScreenTestScene()
         {
@@ -43,17 +45,32 @@ namespace osu.Game.Tests.Visual
                     Name = nameof(ScreenTestScene),
                     RelativeSizeAxes = Axes.Both
                 },
-                content = new Container { RelativeSizeAxes = Axes.Both },
+                new PopoverContainer
+                {
+                    RelativeSizeAxes = Axes.Both,
+                    Children = new Drawable[]
+                    {
+                        content = new Container { RelativeSizeAxes = Axes.Both },
+                        Footer = new ScreenFooter(),
+                    }
+                },
                 overlayContent = new Container
                 {
                     RelativeSizeAxes = Axes.Both,
                     Child = DialogOverlay = new DialogOverlay()
                 },
-                footer = new ScreenFooter(),
             });
 
-            Stack.ScreenPushed += (_, newScreen) => Logger.Log($"{nameof(ScreenTestScene)} screen changed → {newScreen}");
-            Stack.ScreenExited += (_, newScreen) => Logger.Log($"{nameof(ScreenTestScene)} screen changed ← {newScreen}");
+            Stack.ScreenPushed += (oldScreen, newScreen) =>
+            {
+                updateFooter(oldScreen, newScreen);
+                Logger.Log($"{nameof(ScreenTestScene)} screen changed → {newScreen}");
+            };
+            Stack.ScreenExited += (oldScreen, newScreen) =>
+            {
+                updateFooter(oldScreen, newScreen);
+                Logger.Log($"{nameof(ScreenTestScene)} screen changed ← {newScreen}");
+            };
         }
 
         protected void LoadScreen(OsuScreen screen) => Stack.Push(screen);
@@ -77,6 +94,39 @@ namespace osu.Game.Tests.Visual
                 Stack.Exit();
                 return false;
             });
+        }
+
+        private void updateFooter(IScreen? _, IScreen? newScreen)
+        {
+            if (newScreen is OsuScreen osuScreen && osuScreen.ShowFooter)
+            {
+                Footer.Show();
+
+                if (osuScreen.IsLoaded)
+                    updateFooterButtons();
+                else
+                {
+                    // ensure the current buttons are immediately disabled on screen change (so they can't be pressed).
+                    Footer.SetButtons(Array.Empty<ScreenFooterButton>());
+
+                    osuScreen.OnLoadComplete += _ => updateFooterButtons();
+                }
+
+                void updateFooterButtons()
+                {
+                    var buttons = osuScreen.CreateFooterButtons();
+
+                    osuScreen.LoadComponentsAgainstScreenDependencies(buttons);
+
+                    Footer.SetButtons(buttons);
+                    Footer.Show();
+                }
+            }
+            else
+            {
+                Footer.Hide();
+                Footer.SetButtons(Array.Empty<ScreenFooterButton>());
+            }
         }
 
         #region IOverlayManager
